Move existing files to new directory
