+++ opencog/atoms/CMakeLists.txt
@@ -2,6 +2,7 @@ ADD_SUBDIRECTORY (atom_types)
 ADD_SUBDIRECTORY (value)
 ADD_SUBDIRECTORY (base)
 ADD_SUBDIRECTORY (core)
+ADD_SUBDIRECTORY (flow)
 ADD_SUBDIRECTORY (execution)
 ADD_SUBDIRECTORY (pattern)
 ADD_SUBDIRECTORY (reduct)
