+++ opencog/atoms/execution/EvaluationLink.cc
@@ -13,7 +13,7 @@
 #include <opencog/atoms/core/LambdaLink.h>
 #include <opencog/atoms/core/NumberNode.h>
 #include <opencog/atoms/core/PutLink.h>
-#include <opencog/atoms/core/TruthValueOfLink.h>
+#include <opencog/atoms/flow/TruthValueOfLink.h>
 #include <opencog/atoms/execution/Instantiator.h>
 #include <opencog/atoms/pattern/PatternLink.h>
 #include <opencog/atoms/reduct/FoldLink.h>
