+++ opencog/atoms/flow/TruthValueOfLink.cc
