+++ opencog/atoms/flow/TruthValueOfLink.h
@@ -23,7 +23,7 @@
 #ifndef _OPENCOG_TRUTH_VALUE_OF_LINK_H
 #define _OPENCOG_TRUTH_VALUE_OF_LINK_H
 
-#include <opencog/atoms/core/ValueOfLink.h>
+#include <opencog/atoms/flow/ValueOfLink.h>
 
 namespace opencog
 {
