+++ tests/atoms/ValueOfUTest.cxxtest
@@ -7,9 +7,9 @@
  */
 
 #include <opencog/atoms/core/NumberNode.h>
-#include <opencog/atoms/core/ValueOfLink.h>
 #include <opencog/atoms/execution/EvaluationLink.h>
 #include <opencog/atoms/execution/Instantiator.h>
+#include <opencog/atoms/flow/ValueOfLink.h>
 #include <opencog/atoms/reduct/PlusLink.h>
 #include <opencog/atoms/value/FloatValue.h>
 #include <opencog/atomspace/AtomSpace.h>
